requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
gradio>=4.0.0
stability-sdk>=0.8.3
python-dotenv>=1.0.0
pytest>=7.0.0
pillow>=10.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
requests>=2.31.0 